requirements.txt
# ─────────────────────────────────────────────────────────
# Vayu-Rakshak — Python Dependencies
# ─────────────────────────────────────────────────────────

# ── Web Framework ─────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── Database ──────────────────────────────────────────────
sqlalchemy>=2.0.30
pyodbc>=4.0.0
pymssql>=2.2.0

# ── Data & ML ─────────────────────────────────────────────
torch>=2.2.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# ── Dashboard ─────────────────────────────────────────────
streamlit>=1.35.0
streamlit-folium>=0.20.0
folium>=0.16.0
plotly>=5.22.0

# ── AI Agent ──────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.30.0

# ── HTTP / Geo ────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0

# ── Azure Integration ─────────────────────────────────────
azure-storage-blob>=12.19.0
